// src/types/user.ts
export type UserInfoType = {
  displayName: string
  email: string
  businessName: string
  address: string
  phone: string
  dcpLicense: string
  businessEmail: string
  userId: string
}

export type UserType = {
  displayName?: string
  firstName?: string
  lastName?: string
  email: string
  photoUrl?: string
  userId: string
  phoneNumber?: string
  country?: string
  city?: string
  yearOfBirth?: string
  gender?: string
  accessToken?: string
}